requirements.txt
fastapi
uvicorn
python-dotenv
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
langchainhub
langchain-classic
pypdf
supabase>=2.28
httpx>=0.26
websockets>=14
openai
pydantic
langchain